+++ priv/repo/migrations/20260719101804_add_oban_jobs_table.exs
@@ -1,0 +1,7 @@
+defmodule Goodmao2.Repo.Migrations.AddObanJobsTable do
+  use Ecto.Migration
+
+  def up, do: Oban.Migration.up()
+
+  def down, do: Oban.Migration.down(version: 1)
+end